Move resource title under components
