+++ src/components/resource-title/ResourceTitle.scss
