+++ src/opening-period/v2/components/form/OpeningHoursForm.tsx
@@ -31,11 +31,11 @@ import toast from '../../../../components/notification/Toast';
 import OpeningHours from '../opening-hours/OpeningHours';
 import { defaultTimeSpan } from '../../constants';
 import OpeningHoursValidity from './OpeningHoursValidity';
-import ResourceTitle from './ResourceTitle';
 import useMobile from '../../../../hooks/useMobile';
 import { formatDate } from '../../../../common/utils/date-time/format';
 import OpeningHoursTitles from './OpeningHoursTitles';
 import OpeningHoursPreviewMobile from '../preview/OpeningHoursPreviewMobile';
+import ResourceTitle from '../../../../components/resource-title/ResourceTitle';
 
 const getDefaultsValues = (
   datePeriod: DatePeriod | undefined
